+++ stroom-pipeline/src/main/java/stroom/pipeline/destination/RollingFileDestination.java
@@ -19,6 +19,8 @@ package stroom.pipeline.destination;
 import stroom.util.io.ByteCountOutputStream;
 import stroom.util.io.FileUtil;
 import stroom.util.io.PathCreator;
+import stroom.util.io.GZipByteCountOutputStream;
+import stroom.util.io.GZipOutputStream;
 import stroom.util.scheduler.SimpleCron;
 
 import com.google.common.base.Strings;
@@ -27,6 +29,7 @@ import org.slf4j.LoggerFactory;
 
 import java.io.BufferedOutputStream;
 import java.io.IOException;
+import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
@@ -45,6 +48,11 @@ public class RollingFileDestination extends RollingDestination {
     private final Path dir;
     private final Path file;
 
+    /**
+     * Whether to apply GZIP compression to output files
+     */
+    private final boolean useCompression;
+
     public RollingFileDestination(final PathCreator pathCreator,
                                   final String key,
                                   final Long frequency,
@@ -54,7 +62,8 @@ public class RollingFileDestination extends RollingDestination {
                                   final String fileName,
                                   final String rolledFileName,
                                   final Path dir,
-                                  final Path file)
+                                  final Path file,
+                                  final boolean useCompression)
             throws IOException {
         super(key, frequency, schedule, rollSize, creationTime);
 
@@ -65,6 +74,8 @@ public class RollingFileDestination extends RollingDestination {
         this.dir = dir;
         this.file = file;
 
+        this.useCompression = useCompression;
+
         // Make sure we can create this path.
         try {
             if (Files.isRegularFile(file)) {
@@ -73,19 +84,13 @@ public class RollingFileDestination extends RollingDestination {
                 // I have a feeling that the OS might sometimes report that a
                 // file exists that has actually just been rolled.
                 LOGGER.warn("File exists for key={} so rolling immediately", key);
-                setOutputStream(new ByteCountOutputStream(new BufferedOutputStream(Files.newOutputStream(file,
-                        StandardOpenOption.CREATE,
-                        StandardOpenOption.WRITE,
-                        StandardOpenOption.APPEND))));
+                setOutputStream(createOutputStream(file));
 
                 // Roll the file.
                 roll();
 
             } else {
-                setOutputStream(new ByteCountOutputStream(new BufferedOutputStream(Files.newOutputStream(file,
-                        StandardOpenOption.CREATE,
-                        StandardOpenOption.WRITE,
-                        StandardOpenOption.APPEND))));
+                setOutputStream(createOutputStream(file));
             }
         } catch (final IOException e) {
             try {
@@ -94,6 +99,20 @@ public class RollingFileDestination extends RollingDestination {
                 LOGGER.error("Unable to close the output stream.");
             }
             throw e;
+        }
+    }
+
+    private ByteCountOutputStream createOutputStream(final Path file) throws IOException {
+        OutputStream fileOutputStream = Files.newOutputStream(
+                file,
+                StandardOpenOption.CREATE,
+                StandardOpenOption.WRITE,
+                StandardOpenOption.APPEND);
+
+        if (useCompression) {
+            return new GZipByteCountOutputStream(new GZipOutputStream(fileOutputStream));
+        } else {
+            return new ByteCountOutputStream(new BufferedOutputStream(fileOutputStream));
         }
     }