+++ stroom-pipeline/src/main/java/stroom/pipeline/writer/FileAppender.java
@@ -58,6 +58,7 @@ public class FileAppender extends AbstractAppender {
     private final PathCreator pathCreator;
     private ByteCountOutputStream byteCountOutputStream;
     private String[] outputPaths;
+    private boolean useCompression;
 
     @Inject
     public FileAppender(final ErrorReceiverProxy errorReceiverProxy,
@@ -112,8 +113,15 @@ public class FileAppender extends AbstractAppender {
             LOGGER.trace("Creating output stream for path {}", path);
 
             // Get a writer for the new lock file.
-            byteCountOutputStream = new ByteCountOutputStream(
-                    new BufferedOutputStream(Files.newOutputStream(lockFile)));
+            if (useCompression) {
+                byteCountOutputStream =
+                        new GZipByteCountOutputStream(new GZipOutputStream(Files.newOutputStream(lockFile)));
+            }
+            else {
+                byteCountOutputStream =
+                        new ByteCountOutputStream(new BufferedOutputStream(Files.newOutputStream(lockFile)));
+            }
+
             return new LockedOutputStream(byteCountOutputStream, lockFile, outFile);
 
         } catch (final IOException e) {
@@ -141,6 +149,9 @@ public class FileAppender extends AbstractAppender {
     public void setOutputPaths(final String outputPaths) {
         this.outputPaths = outputPaths.split(",");
     }
+
+    @PipelineProperty(description = "Apply GZIP compression to output files", defaultValue = "false")
+    public void setUseCompression(final boolean useCompression) { this.useCompression = useCompression; }
 
     @SuppressWarnings("unused")
     @PipelineProperty(